Rename : Header.js and Params.js to HeaderManager and ParamManager
